Split reader component
