+++ client/src/routes/reader.$bookId.tsx
@@ -1,35 +1,5 @@
-import {
-  Comment,
-  CommentOutlined,
-  EmojiEmotions,
-  MoreVert,
-  Note,
-  NoteAdd,
-  Send,
-} from "@mui/icons-material";
-import {
-  Badge,
-  Box,
-  Button,
-  Card,
-  CardActions,
-  CardContent,
-  CardHeader,
-  Divider,
-  Drawer,
-  Fab,
-  IconButton,
-  Input,
-  InputAdornment,
-  Modal,
-  Popover,
-  Stack,
-  Typography,
-  useTheme,
-  MenuItem,
-  Menu,
-  Avatar,
-} from "@mui/material";
+import { Note, NoteAdd } from "@mui/icons-material";
+import { Badge, Box, Drawer, Fab, Stack, useTheme } from "@mui/material";
 import { createFileRoute } from "@tanstack/react-router";
 import { EpubCFI, Rendition } from "epubjs";
 import { useEffect, useMemo, useRef, useState } from "react";
@@ -41,9 +11,8 @@ import useAutoLogin from "../api/login/useAutoLogin";
 import useAutoTokenRefresh from "../api/login/useAutoTokenRefresh";
 import loadBook from "../util/loadBook";
 import useInvalidateQueriesOnAuthChange from "../api/login/useInvalidateQueriesOnAuthChange";
-import State from "../states";
-import { useAtomValue } from "jotai";
-import type { HighlightReactionType } from "../types/highlight";
+import HighlightCard from "../component/HighlightCard";
+import HighlightCreationModal from "../component/HighlightCreationModal";
 
 export const Route = createFileRoute("/reader/$bookId")({
   component: RouteComponent,
@@ -338,281 +307,4 @@ function diffMemos(prev: Highlight[], next: Highlight[]): HighlightDiff {
     added,
     removed,
   };
-}
-
-function HighlightCard({
-  highlight,
-  refetchHighlights,
-  activityId,
-}: {
-  highlight: Highlight;
-  activityId?: number;
-  refetchHighlights: () => void;
-}) {
-  const user = useAtomValue(State.Auth.user);
-  const [openComments, setOpenComments] = useState(false);
-  const [commentContent, setCommentContent] = useState("");
-  const isAuthor = highlight.authorId === user?.memberId;
-  const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
-  const [emojiAnchorEl, setEmojiAnchorEl] = useState<null | HTMLElement>(null);
-  const emojiList: { type: HighlightReactionType; emoji: string }[] = [
-    { type: "GREAT", emoji: "👍" },
-    { type: "HEART", emoji: "❤️" },
-    { type: "SMILE", emoji: "😊" },
-    { type: "CLAP", emoji: "👏" },
-    { type: "SAD", emoji: "😢" },
-    { type: "ANGRY", emoji: "😡" },
-    { type: "SURPRISED", emoji: "😲" },
-  ];
-
-  const { data: reactions, refetch: refetchReactions } = useQuery({
-    queryKey: ["highlightComments", highlight.id],
-    queryFn: async () => {
-      const response =
-        await API_CLIENT.highlightController.getHighlightReactions(
-          highlight.id
-        );
-      if (!response.isSuccessful) {
-        throw new Error(response.errorMessage);
-      }
-      const reactions: Map<
-        HighlightReactionType,
-        (typeof response.data)[number][]
-      > = new Map([
-        ["GREAT", []],
-        ["HEART", []],
-        ["SMILE", []],
-        ["CLAP", []],
-        ["SAD", []],
-        ["ANGRY", []],
-        ["SURPRISED", []],
-      ]);
-      response.data.forEach((reaction) => {
-        reactions.get(reaction.reactionType!)!.push(reaction);
-      });
-      console.log(reactions);
-      return reactions;
-    },
-    placeholderData: keepPreviousData,
-  });
-
-  const onReactionClicked = async (reaction: HighlightReactionType) => {
-    const response = await API_CLIENT.reactionController.addReaction(
-      highlight.id,
-      {
-        reactionType: reaction,
-      }
-    );
-    if (!response.isSuccessful) {
-      alert(response.errorMessage);
-    }
-    setEmojiAnchorEl(null);
-    refetchReactions();
-  };
-
-  const onShareToGroupClicked = async () => {
-    const response = await API_CLIENT.highlightController.updateHighlight(
-      highlight.id,
-      {
-        activityId,
-      }
-    );
-    if (!response.isSuccessful) {
-      alert(response.errorMessage);
-    }
-    setAnchorEl(null);
-    refetchHighlights();
-  };
-
-  return (
-    <Card>
-      <Menu
-        anchorEl={anchorEl}
-        open={!!anchorEl}
-        onClose={() => setAnchorEl(null)}
-      >
-        <MenuItem
-          disabled={!!highlight.activityId}
-          onClick={onShareToGroupClicked}
-          value="group"
-        >
-          그룹 공개
-        </MenuItem>
-      </Menu>
-      <CardContent sx={{ pt: 1 }}>
-        <Stack spacing={1}>
-          <Stack spacing={1} direction={"row"} alignItems={"center"}>
-            <Avatar src={highlight.authorProfileImageURL} />
-            <Typography variant="body1">{highlight.authorName}</Typography>
-            {isAuthor && (
-              <IconButton
-                onClick={(e) => {
-                  setAnchorEl(e.currentTarget);
-                }}
-              >
-                <MoreVert />
-              </IconButton>
-            )}
-          </Stack>
-          <Typography variant="body1">{highlight.memo}</Typography>
-          {reactions && (
-            <Stack direction="row" spacing={1} alignItems="center">
-              {emojiList.map((e) => {
-                const count = reactions.get(e.type)?.length || 0;
-                if (count === 0) return null;
-                return (
-                  <Box
-                    key={e.type}
-                    sx={{ display: "flex", alignItems: "center", gap: 0.5 }}
-                  >
-                    <span style={{ fontSize: 20 }}>{e.emoji}</span>
-                    <Typography variant="body2" color="text.secondary">
-                      {count}
-                    </Typography>
-                  </Box>
-                );
-              })}
-            </Stack>
-          )}
-        </Stack>
-      </CardContent>
-      <CardActions sx={{ justifyContent: "flex-end" }}>
-        <IconButton size="small" onClick={() => setOpenComments(!openComments)}>
-          <Badge badgeContent={2} color="primary">
-            {openComments ? <Comment /> : <CommentOutlined />}
-          </Badge>
-        </IconButton>
-        <IconButton
-          size="small"
-          onClick={(e) => setEmojiAnchorEl(e.currentTarget)}
-        >
-          <EmojiEmotions />
-        </IconButton>
-        <Popover
-          open={Boolean(emojiAnchorEl)}
-          anchorEl={emojiAnchorEl}
-          onClose={() => setEmojiAnchorEl(null)}
-          anchorOrigin={{ vertical: "top", horizontal: "center" }}
-          transformOrigin={{ vertical: "bottom", horizontal: "center" }}
-        >
-          <Stack direction="row" spacing={1}>
-            {emojiList.map((e) => (
-              <IconButton
-                key={e.type}
-                onClick={() => {
-                  onReactionClicked(e.type);
-                }}
-                size="medium"
-              >
-                {e.emoji}
-              </IconButton>
-            ))}
-          </Stack>
-        </Popover>
-      </CardActions>
-      {openComments && (
-        <>
-          {new Array(2).fill(0).map(() => {
-            return (
-              <>
-                <Divider />
-                <CardHeader title="Nickname2" avatar="2" sx={{ pb: 0 }} />
-                <CardContent sx={{ pt: 1 }}>
-                  <Typography variant="body1">Comment</Typography>
-                </CardContent>
-              </>
-            );
-          })}
-          <Divider />
-          <CardActions>
-            <Input
-              value={commentContent}
-              onChange={(e) => setCommentContent(e.target.value)}
-              fullWidth
-              multiline
-              size="small"
-              endAdornment={
-                <InputAdornment position="end">
-                  <IconButton edge="end">
-                    <Send />
-                  </IconButton>
-                </InputAdornment>
-              }
-            />
-          </CardActions>
-        </>
-      )}
-    </Card>
-  );
-}
-
-function HighlightCreationModal({
-  open,
-  onClose,
-  selection,
-  addHighlight,
-}: {
-  open: boolean;
-  onClose: () => void;
-  selection: Selection | null;
-  addHighlight: (props: { memo: string; cfi: string }) => void;
-}) {
-  const [content, setContent] = useState("");
-
-  if (!selection) {
-    return null;
-  }
-
-  return (
-    <Modal open={open} onClose={onClose}>
-      <Box
-        sx={{
-          position: "absolute",
-          top: "50%",
-          left: "50%",
-          transform: "translate(-50%, -50%)",
-          width: 400,
-        }}
-      >
-        <Card>
-          <CardHeader title="Create Memo" />
-          <CardContent>
-            <Typography color="textSecondary" variant="body1">
-              {selection.text}
-            </Typography>
-            <Divider />
-            <Input
-              value={content}
-              onChange={(e) => setContent(e.target.value)}
-              type="textarea"
-              multiline
-              fullWidth
-            />
-          </CardContent>
-          <CardActions>
-            <Button color="secondary" onClick={onClose}>
-              Cancel
-            </Button>
-            <Button
-              variant="contained"
-              color="primary"
-              onClick={() => {
-                if (!selection) {
-                  return;
-                }
-                addHighlight({
-                  memo: content,
-                  cfi: selection.epubcfi,
-                });
-                setContent("");
-                onClose();
-              }}
-            >
-              Create
-            </Button>
-          </CardActions>
-        </Card>
-      </Box>
-    </Modal>
-  );
 }
